+++ Runtime/Controllers/RateManagerUpdateRateController.cs
@@ -2,6 +2,14 @@ using System;
 using UnityEngine;
 using UniRate.Debug;
 
+#if UNITY_2021_1_OR_NEWER
+using UnityApplication = UnityEngine.Device.Application;
+using UnityScreen = UnityEngine.Device.Screen;
+#else
+using UnityApplication = UnityEngine.Application;
+using UnityScreen = UnityEngine.Screen;
+#endif
+
 namespace UniRate.Internals {
 
     public class RateManagerUpdateRateController : RateManagerValueController {
@@ -116,7 +124,7 @@ namespace UniRate.Internals {
                     // QualitySettings.vSyncCount value must be 0, 1, 2, 3, or 4.
                     // QualitySettings.vSyncCount is ignored on iOS.
                     vSyncCount = Mathf.Clamp(
-                        Screen.currentResolution.refreshRate / targetFrameRate,
+                        UnityScreen.currentResolution.refreshRate / targetFrameRate,
                         1,
                         4
                     );
@@ -132,11 +140,11 @@ namespace UniRate.Internals {
                 QualitySettings.vSyncCount = vSyncCount;
             }
 
-            if (Application.targetFrameRate != targetFrameRate) {
+            if (UnityApplication.targetFrameRate != targetFrameRate) {
                 if (RateDebug.IsLogLevelActive(RateLogLevel.Trace)) {
                     RateDebug.Log(RateLogLevel.Trace, $"setting Application.targetFrameRate to {targetFrameRate.ToString()}");
                 }
-                Application.targetFrameRate = targetFrameRate;
+                UnityApplication.targetFrameRate = targetFrameRate;
             }
         }